refactor(main): move run() to commands/mod.rs, trim main.rs (step 4)
Final step of #127. main.rs now contains only main() + module
declarations (26 lines, under the 100-line acceptance criterion).

  src/util.rs            new — holds get_filename, the shared
                         path-to-display-name helper used by both
                         commands/* and processors/*
  src/commands/mod.rs    gains pub fn run(opts) — the validation
                         and dispatch logic that previously lived
                         in main.rs

main.rs is now: imports + main(), nothing else. All command and
processing logic is owned by the appropriate module tree.

No public library API change. CLI behavior unchanged.
