+++ src/commands/apply.rs
@@ -5,10 +5,10 @@ use std::path::PathBuf;
 
 use crate::cli::options::{Options, OutputFormat};
 use crate::commands::utils::{create_json_summary, print_dry_run_notice, update_counters};
-use crate::get_filename;
 use crate::json_output::{JsonFileResult, JsonOutput};
 use crate::processors::apply::{process_apply, process_apply_channel};
 use crate::progress::{create_progress_bar, progress_finish, progress_inc, progress_set_message};
+use crate::util::get_filename;
 
 pub fn cmd_apply(files: &[PathBuf], steps: i32, opts: &Options) -> Result<()> {
     if steps == 0 {
