+++ src/commands/undo.rs
@@ -4,10 +4,10 @@ use std::path::PathBuf;
 
 use crate::cli::options::{Options, OutputFormat};
 use crate::commands::utils::{create_json_summary, print_dry_run_notice, update_counters};
-use crate::get_filename;
 use crate::json_output::{JsonFileResult, JsonOutput};
 use crate::processors::undo::process_undo;
 use crate::progress::{create_progress_bar, progress_finish, progress_inc, progress_set_message};
+use crate::util::get_filename;
 
 pub fn cmd_undo(files: &[PathBuf], opts: &Options) -> Result<()> {
     let dry_run_prefix = opts.dry_run_prefix();
